requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
click==7.1.2
colorama==0.4.4
flake8==3.8.4
Flask==1.1.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.8
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.1
pytest-flake8==1.0.7
python-dotenv==0.15.0
toml==0.10.2
Werkzeug==1.0.1
